Add some missing patches to the script compiler

+++ neo/game/script/Script_Program.h
@@ -4,7 +4,7 @@
 Doom 3 GPL Source Code
 Copyright (C) 1999-2011 id Software LLC, a ZeniMax Media company. 
 
-This file is part of the Doom 3 GPL Source Code ("Doom 3 Source Code").  
+This file is part of the Doom 3 GPL Source Code ("Doom 3 Source Code").
 
 Doom 3 Source Code is free software: you can redistribute it and/or modify
 it under the terms of the GNU General Public License as published by
@@ -96,7 +96,7 @@ class idTypeDef {
 private:
 	etype_t						type;
 	idStr 						name;
-	int							size;
+	size_t						size;
 
 	// function types are more complex
 	idTypeDef					*auxType;					// return type
@@ -442,7 +442,7 @@ private:
 	idStr 										filename;
 	int											filenum;
 
-	int											numVariables;
+	size_t											numVariables;
 	byte										variables[ MAX_GLOBALS ];
 	idStaticList<byte,MAX_GLOBALS>				variableDefaults;
 	idStaticList<function_t,MAX_FUNCS>			functions;
